pyproject.toml
[tool.poetry]
name = "adam"
version = "0.1.0"
description = ""
authors = ["Shannon Gray <shanngray@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.1"
python-dotenv = "^1.0.1"
langgraph = "^0.0.62"
langchain-experimental = "^0.0.59"
langchain-cohere = "^0.1.5"
langchain-core = "^0.2.3"
langchain-text-splitters = "^0.2.0"
langchain-groq = "^0.1.4"
fastapi = "^0.112.2"
sse-starlette = "^2.1.3"
asyncio = "^3.4.3"
httpx = "^0.27.2"
aiosqlite = "^0.20.0"
icecream = "^2.1.3"
greenlet = "^3.0.3"
websockets = "^13.0.1"
wsproto = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
